requirements.txt
Flask
transformers
Pillow
torch
